Moved all entities and extensions to the separate files.
